Fix peer selection for gossip
The random algorithm to select a peer does not work properly for smaller
sets of nodes (like 3), where sometimes it returns the same node to gossip
with for 5+ consecutive time.
If a similar streak occurs at other nodes, there will a situation where
nodes gossip with only each other in silos declaring other nodes to be dead.
To avoid such situations, gossip now selects peers in a round robin way.
It ensures that we gossip with all peers.
To make sure not all nodes try to gossip with the same node (like all of them
first gossip with node 0, then node 1) which would be a problem, the gossip
algorithm selects a peer to start who is next to it if ordered by name.