1) Fix the issue where a node which never gossips is never marked down To handle this, we need to add node id of all the nodes that get added. The timeout to mark this or any node dead when gossip begins is 4 times the set death interval. This is to handle the case where nodes usually take more time when booting up.
2) Add a generation number with gossip.
This is used to handle the case where gossip marks a node down because
of a stale update. This can happen when:
1. node goes down
2. gossip determines node is down and starts spreading the message
3. node comes back again.

Other nodes which get this stale message in (2) may already be communicating
via other means with node 2. In such a scenario, they can ask gossip
to mark its information for this node to be of old generation.
Gossip will then wait for an update for this node, and once it arrives,
it will mark the nodes status appropriately.
If the update never arrives, this node is marked down with a different
down status in order to note the fact that it was waiting for an update which
never came.

Added an API to mark the node info belonging to old generation, and two
new node status - one waiting for update, and another, down waiting for update.