Don't allow nodes with different versions to communicate.
Gossip v1 had the ability to ignore messages from nodes with other versions.
But at the same time it use to ignore this check if the node was already in its map.

Gossip v2 added the ability to not allow nodes with different versions to talk
even if they are present in node map. It overcame the bug in v1 by allowing the v2
nodes to start gossiping only to the nodes with same version.

However memberlist displays a wierd behavior wherein if a node sends a Alive message
it starts replying back with an Alive message ignoring our version check.

This fix overcomes the above memberlist limitation and works with v1 nodes as well.